ipc: sysv shared memory limited to 8TiB

+++ include/linux/ipc_namespace.h
@@ -43,7 +43,7 @@ struct ipc_namespace {
 	size_t		shm_ctlmax;
 	size_t		shm_ctlall;
 	int		shm_ctlmni;
-	int		shm_tot;
+	unsigned long	shm_tot;
 	/*
 	 * Defines whether IPC_RMID is forced for _all_ shm segments regardless
 	 * of shmctl()
